feat: Change design for 404 page
