+++ src/app/app.routes.ts
@@ -1,6 +1,6 @@
 import { Routes } from '@angular/router';
 import { MainComponent } from './pages/main.component';
-import { NotFoundComponent } from './pages/404-page.component';
+import { NotFoundComponent } from './pages/404-page/404-page.component';
 import { LoginComponent } from './pages/login.component';
 import { RegistrationComponent } from './pages/registration/registration.component';
 import { HeroComponent } from './pages/hero.component';
