+++ src/app/components/header/header.component.scss
@@ -23,6 +23,13 @@
 
 .logo {
   width: 50px;
+  transition: ease-in-out 0.5s;
+
+  @include hover {
+    &:hover {
+      transform: scale(1.1);
+    }
+  }
 }
 
 .main-bar {
