+++ src/app/pages/404-page/404-page.component.html
@@ -1,0 +1,18 @@
+<div class="not-found-wrapper">
+  <h3>404</h3>
+  <div class="paragraph-wrapper">
+    <p>Oops! This page does not exist.</p>
+    <p>
+      <a
+        routerLink="/main"
+        routerLinkActive="active"
+        ariaCurrentWhenActive="page"
+        >Return to homepage</a
+      >
+    </p>
+    <p>
+      You will be redirected automatically in
+      <span>{{ counter() }}</span> seconds...
+    </p>
+  </div>
+</div>
