+++ src/app/pages/404-page/404-page.component.scss
@@ -1,0 +1,50 @@
+@import "../../../styles/colors";
+@import "../../../styles/breakpoints";
+
+::ng-deep app-not-found {
+  height: calc(100svh - 80px);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+.not-found-wrapper {
+  max-width: 50vw;
+  padding: 5px;
+  display: flex;
+  text-align: center;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+
+  h3 {
+    font-size: 8em;
+    font-weight: 600;
+    color: $symbols;
+    margin-bottom: 1em;
+  }
+}
+.paragraph-wrapper {
+  p {
+    padding: 10px;
+    font-size: 1.5em;
+    font-weight: 400;
+    color: $text;
+
+    span {
+      font-weight: 600;
+      color: $symbols;
+    }
+  }
+  a {
+    font-size: 2em;
+    color: $header-yellow;
+    text-decoration: none;
+    font-weight: bold;
+
+    @include hover {
+      &:hover {
+        text-decoration: underline;
+      }
+    }
+  }
+}
